choice field: extract hooks from BaseDynamicChoiceField
