+++ packages/admin/src/components/bindingFacade/collections/DataGrid/cells/SelectCellFilter.tsx
@@ -3,7 +3,7 @@ import { useMemo } from 'react'
 import { useMessageFormatter } from '../../../../../i18n'
 import { ChoiceFieldData, MultiSelectFieldInner } from '../../../fields'
 import { BaseDynamicChoiceField } from '../../../fields/ChoiceField/BaseDynamicChoiceField'
-import { useSelectOptions } from '../../../fields/ChoiceField/useSelectOptions'
+import { useSelectOptions } from '../../../fields/ChoiceField/hooks/useSelectOptions'
 import { FilterRendererProps } from '../base'
 import { dataGridCellsDictionary } from './dataGridCellsDictionary'
 import { EntityAccessor, EntityId } from '@contember/binding'
