+++ packages/admin/src/components/bindingFacade/fields/ChoiceField/hooks/useCurrentValues.ts
@@ -1,0 +1,12 @@
+import { useDesugaredOptionPath } from './useDesugaredOptionPath'
+import { useNormalizedOptions } from './useNormalizedOptions'
+import { BaseDynamicChoiceField } from '../BaseDynamicChoiceField'
+import { EntityAccessor } from '@contember/binding'
+
+export const useCurrentValues = (
+	optionProps: BaseDynamicChoiceField,
+	currentlyChosenEntities: EntityAccessor[],
+) => {
+	const desugaredOptionPath = useDesugaredOptionPath(optionProps)
+	return useNormalizedOptions(currentlyChosenEntities, desugaredOptionPath, optionProps)
+}
