+++ packages/admin/src/components/bindingFacade/fields/ChoiceField/hooks/useDynamicMultipleChoiceWithConnectingEntityField.ts
@@ -7,12 +7,12 @@ import {
 	useSortedEntities,
 } from '@contember/binding'
 import { useCallback, useMemo } from 'react'
-import { useCurrentValues } from './BaseDynamicChoiceField'
-import type { ChoiceFieldData } from './ChoiceFieldData'
+import type { ChoiceFieldData } from '../ChoiceFieldData'
 import { useSelectOptions } from './useSelectOptions'
-import { useAccessorErrors } from '../../errors'
+import { useAccessorErrors } from '../../../errors'
 import { useOnAddNew } from './useOnAddNew'
 import { DynamicMultipleChoiceFieldProps } from './useDynamicMultipleChoiceField'
+import { useCurrentValues } from './useCurrentValues'
 
 export interface DynamicMultipleChoiceWithConnectingEntityFieldProps {
 	connectingEntityField: string | SugaredRelativeSingleEntity
@@ -38,9 +38,9 @@ export const useDynamicMultipleChoiceWithConnectingEntityField = (
 	}, [optionTargetField, sortedConnectingEntities.entities])
 
 
-	const options = useSelectOptions(props, currentlyChosenEntities)
+	const options = useSelectOptions(props, currentlyChosenOptions)
 
-	const currentValues = useCurrentValues(currentlyChosenOptions, props)
+	const currentValues = useCurrentValues(props, currentlyChosenOptions)
 
 	const getConnectingEntityValues = connectingEntitiesListAccessor.getAccessor
