+++ packages/admin/src/components/bindingFacade/fields/ChoiceField/index.ts
@@ -3,5 +3,5 @@ export * from './choiceFieldDictionary'
 export * from './ChoiceFieldData'
 export * from './DynamicMultiChoiceField'
 export * from './StaticSingleChoiceField'
-export * from './useDynamicMultipleChoiceField'
-export * from './useDynamicSingleChoiceField'
+export * from './hooks/useDynamicMultipleChoiceField'
+export * from './hooks/useDynamicSingleChoiceField'
